+++ src/plugins/intel_cpu/src/nodes/if.cpp
@@ -249,7 +249,7 @@ void If::prepareAfterMappers(const bool isThen, const dnnl::engine& eng) {
 
 std::deque<MemoryPtr> If::getToMemories(const Node* node, const size_t port) {
     std::deque<MemoryPtr> memories;
-    for (const auto& edge : node->getChildEdgesAtPort(port)) {
+    for (const auto& edge : node->getChildEdgesAtPort(static_cast<int>(port))) {
         memories.push_back(edge->getMemoryPtr());
     }
     return memories;
